requirements.txt
# Core API Framework
fastapi==0.111.0
uvicorn[standard]==0.29.0
python-multipart==0.0.9

# AI / ML
tensorflow-macos
tensorflow-metal
numpy==1.26.4
Pillow==10.3.0
scikit-learn==1.4.2

# Database
motor==3.4.0           # Async MongoDB driver
pymongo==4.7.2

# Utilities
python-dotenv==1.0.1
pydantic==2.7.1
pydantic-settings==2.2.1
aiofiles==23.2.1
httpx==0.27.0

# Testing
pytest==8.2.0
pytest-asyncio==0.23.6
httpx==0.27.0
